GHA: spellcheck
This spellchecker checks markdown files. For this reason this job
converts all man pages in the repository to markdown with pandoc before
the check runs. I am unfortunately not aware of any tool that can run
spellcheck on the "raw" nroff file.

This is somewhat of an experiment. I am not sure this will make total
sense as maybe the amount of unknown words will be too large and not
manageable. But I do know that I write a lot of typos and by checking
our documentation with tools we might improve it.